Fixed AddImage button to work in AddPin t#118
